Fix banned account swaps
When a swap is from a banned wallet, the collect button is currently removed in the swap list of an objkt, so no-one can collect from the banned account. But if the banned wallet is the first swap, then the swap is still available through the main collect button.

This fix totally remove swaps from banned account, including on the main collect button


Former-commit-id: 754f4bcbf03b0360635d7d2c84857c616a650850